% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgres-package.R
\docType{package}
\name{pgres-package}
\alias{pgres}
\alias{pgres-package}
\title{pgres: PostgreSQL Utility Helper Functions in R}
\description{
Suite of Database Utility Helpers Specific to PostgreSQL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/pgres}
  \item \url{https://jimbrig.github.io/pgres/}
  \item Report bugs at \url{https://github.com/jimbrig/pgres/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
