% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_primary_key.R
\name{set_foreign_key}
\alias{set_foreign_key}
\title{Set Foreign Key}
\usage{
set_foreign_key(
  conn,
  table,
  key,
  foreign_table,
  foreign_key = key,
  constraint = paste(table, key, "fk", sep = "_")
)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Table to add FK constraint to}

\item{key}{Column in \code{table} to act as the \code{key} referencing \code{foreign_key}}

\item{foreign_table}{Table to reference}

\item{foreign_key}{Column to reference from \code{foreign_table}}

\item{constraint}{Name of the constraint to be created.}
}
\value{
SQL query
}
\description{
Creates a \verb{FOREIGN KEY} \code{REFERENCES} constraint on the specified
\code{key} referencing the \code{foreign_table}'s \code{foreign_key} with the constaint
named using the \code{constraint} argument.

The SQL query utilized is as follows:\if{html}{\out{<div class="sourceCode sql">}}\preformatted{ALTER TABLE <schema>.<table>
ADD CONSTRAINT <constraint> FOREIGN KEY (<key>)
REFERENCES <foreign_table> (<foreign_key>);
}\if{html}{\out{</div>}}
}
