% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_primary_key.R
\name{set_primary_key}
\alias{set_primary_key}
\title{Set Primary Key}
\usage{
set_primary_key(conn = getOption(pg.conn, pg_connect()), table, column)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Character string representing the table (with optional schema
prefix) to alter.}

\item{column}{Character string representing the column(s) to add as primary
keys. Note that multiple columns are allowed to create a \emph{conjugate primary
key}.}
}
\value{
returns the SQL query for reference
}
\description{
Creates a \verb{PRIMARY KEY} constraint on the specified \code{column}
for the provided \code{table} via the SQL query:
\verb{ALTER TABLE <schema>.<table> ADD PRIMARY KEY (<column>);}.
}
\examples{
\dontrun{
conn <- pg_connect()

# set a primary key on the table 'users' for column 'user_id':
set_primary_key(conn, "users", "user_id")

}

}
